Remove broken C-struct namespace.
The idea was to automatically prefix structs with a "module___" string
to emulate a namespace in C. However, it is not so easy because this
change also needs to be applied to nested uses of the struct. Remove
for now.

Consequently, it is up to the user to choose sane struct names that
include a namespace (e.g. struct kdl_twist opposed to struct
twist). It should not be a problem.