ubx_launch: add start directive.
The motivation is to allow systemmodels to not only instantiate,
configure and connect blocks, but also to startup the blocks such that
the system can start to run.

To that end a new "start" keyword is introduced, that allows to define
a list of "last to be started" blocks, which are typically the
triggers:

   start = { "trigger1", "trigger2", ... }

This will lead to all other blocks to be activated before the ones in
the start list. After that, the blocks in the start list will be
activated in the given order.

See also the examples trig_rnd_hexdump.[usc|json].

For testing purposes, the starting can be also be disabled by using
the "-nostart" cmdline switch to ubx_launch.