config data is not presized to 1
This used to be a useful default when almost any config was of size
one. Since more and more dynamically sized configs are used, it
becomes more cumbersome to distinguish between unconfigured
configurations and legal ones of size 1.

This patch disables the presizing, so by default configurations are of
size zero. This way, configurations which have not been assigned will
be of len=0. Consequently, block code must handle the case that
ubx_config_get_data_ptr may return NULL and len=0 and (if applicable)
illegal non-zero values of len.