ubx.lua: load unversioned libubx.so
Using the versioned libubx.so.X is a pain to update with each release
and is currently not really working anyway, as the header files
included by luajit are not versioned.

This could be fixed rather easily, however as most people are not
using multiple versions of microblx simultaneously, drop this for now.