rtlog: add block loglevel handling
This adds support for controlling the loglevel on a per block
basis. This is achieved by means of an 'int' block configuration
'loglevel', which the user is responsible for creating. If it exists
at init time, it will be used as the current block loglevel.

If no loglevel configuration is found, then all log messages are
forwarded and will be filtered by the node loglevel.

Signed-off-by: Markus Klotzbuecher <mk@mkio.de>