Use inotify to detect availability of shm file
The logging client relies on the availability of a shm file in order to
be able to read messages logged to the rtlogger subsystem. If the logging
client is launched before the function blocks are launched, the shm file
will not be available, and hence the logging client will terminate.

With this patch we attempt to open the shm file and if it does not exist we
utilise inotify to notify us when the shm file becomes available, and on
becoming available, we open the newly available shm file and log from there.

We also receive inotify events while the logger is running and if the function
blocks are pulled down, the logger will continue to log until it's read pointer
catches up to the log aggregator's write pointer. When the function blocks are
re-launched, we recieve an inotify event which tells us there is a new shm
file available and we close the old one, open the new one and carry on logging.

Signed-off-by: Hamish Guthrie <hamish.guthrie@kistler.com>