blockdiagram: correct check for non-existing configs
In luajit checking a cdata for being NULL can not be done by negating
it (it is a boxed value, so that evaluates to false). Instead, it must
be explicitely compared against nil.

This problem was previously hidden because block_config_get would
throw an error if a config of the given name didn't exist. Since that
function was updated to just return NULL this problem could occur.

Reported-by: Hamish Guthrie <hamish.guthrie@kistler.com>
Signed-off-by: Markus Klotzbuecher <mk@mkio.de>