Add additional command-line arguments
Added additional command-line arguments to be able to specify
source file, target file and array name, this permits the tool
to be used where the source file could be a header file in a
library for example. Thus we can specify the source as well as
the target.

If the additional command-line arguments are not used, the
behaviour remains the same as previously

Signed-off-by: Hamish Guthrie <hamish.guthrie@kistler.com>