blockdiagram: drop start directive and handle sequence automatically
This patch removes the the blockdiagram start directive, which was
used to indicate the blocks to be started in the end of the
startup (e.g. ptrig and similar blocks). Since we now have the
BLOCK_ATTR_ACTIVE attribute available, we don't require this
information anymore and can just startup all active blocks after all
others.

Signed-off-by: Markus Klotzbuecher <mk@mkio.de>