blockdiagram: ubx_launch add support for model merging
Add support for merging models. This can be used to extend models at
launch time without polluting their specification beforehand. The main
use-case is for combining an activity model (e.g. a ptrig block and
configuration) with a reusable model without polluting the latter with
specifics such thread period and priority.

As a first example, this patch refactors the deep_composition usc to
use passive trigs and moves the active ptrig to a separate ptrig.usc
file.

Now the (reusable) deep_composition is not polluted with activity
related aspects and can be reused stand alone. In order to introduce
the activity, the ptrig.usc model can be merged by providing it as a
second argument to the `-c` option:

$ cd examples/systemmodels/composition/
$ ubx_launch -c deep_composition.usc,ptrig.usc

Signed-off-by: Markus Klotzbuecher <mk@mkio.de>