ubx_launch: allow loading blocks from multiple prefixes
With the current implementation, the first prefix in UBX_PATH in which
the core files are found is selected and ubx core and blocks are
loaded from it. This changes the behavior to allow blocks to be loaded
from any of the prefixes contained in UBX_PATH.

However with this change, the UBX_PATH prefixes are not added to the
list of default paths anymore, hence if used, UBX_PATH must include
the default paths too.

Signed-off-by: Markus Klotzbuecher <mk@mkio.de>