lfds_cyclic: add loglevel_overruns config
Not all overruns indicate an error, depending on the architecture this
can be legal. This adds a config loglevel_overruns to the lfds_cyclic
iblock that permits configuring the loglevel of the message output
upon overrun. If -1, logging overruns is disabled. The default is kept
at the previous level UBX_LOGLEVEL_NOTICE.

Signed-off-by: Markus Klotzbuecher <mk@mkio.de>