core: config_get_data_ptr: only update ptr if len>0
In ubx_config_get_data_ptr the out value would be set even if the
config in question was unconfigured. Although the len return value is
normally used to check whether a value exists, this leaves the ptr
pointing to some seemingly valid (realloced) but potentially
non-existing memory.

This fixes the function to only set the out ptr if the config length
is larger than 0 (i.e. has been configured to something).

Signed-off-by: Markus Klotzbuecher <mk@mkio.de>