core: ubx_node_cleanup: don't crash on unregistered blocks
When unregistering a block fails due to a wrong name being used, the
block memory would still be unmapped by unloading the module and
checking for the number of blocks using HASH_COUNT would crash because
a block in unmapped memory is accessed. This splits the module unload
into cleanup and close and does the check in between.

Signed-off-by: Markus Klotzbuecher <mk@mkio.de>