core: ubx_block_start: set block state active before running start
This avoids a race condition in which a high priority pthread preempts
the called of block_start before it could set the new block state. In
that case it remains in inactive until the called runs again. In the
case of autostop_steps this causes problems due to ubx_block_stop
failing because the ptrig block is still in inactive.

Signed-off-by: Markus Klotzbuecher <mk@mkio.de>