core, make: remove cproto dependency
ubx_proto.h header is split into several smaller ones which shall be
maintained by hand in the future. If convenient, cproto can still be
used a a dev tool, but it's not a build dependency anymore. By this,
the cproto error about the missing expected symbol in rtlog is removed
too.

Signed-off-by: Markus Klotzbuecher <mk@mkio.de>