lfds_cyclic: add allow_partial configuration
This concerns the behavior of lfds_cyclic for the case when a message
with an array len < than the configured data_len. The previous
behavior was to just take the partial message, but this may (or may
not) be silently ignoring a bug. To resolve this, the default of
lfds_cyclic is to reject and complain about partial messages, unless
the config allow_partial is explicitely set to 1.

Signed-off-by: Markus Klotzbuecher <mk@mkio.de>