trigger: add multichain support
This is a large change to introduce support for multiple trigger
chains for trig and ptrig. Per default still one chain exists (config
name is "chain0"). Depending on the value of the config "num_chains",
additional configs called "chainX" will be added to the block
interface.

As a consequence, chains now need to be configured in in state
inactive opposed to previously init. The default chain initially
triggered is chain0. To switch chains write a long value to the port
"active_chain".

In the course of this change, some cosmetic renaming was done:

 struct trig_info -> ubx_chain
 struct ubx_trig_spec -> ubx_triggee

in trig_utils:
 trig_info_* -> ubx_chain_*

Signed-off-by: Markus Klotzbuecher <mk@mkio.de>