mqueue, ubx-mq: fix issue caused by strictler cleanup
Using a strictler block cleanup policy caused the mqueues created by
ubx-mq tool to be properly destructed, but in their cleanup they
called mq_unlink which removed the queue. This adds a mqueue config
`unlink` to defined whether the queue shall be removed or not (default
yes).

In addition, the ubx.node_cleanup is set as the default gc method for
node instead of ubx.node_rm. This leaves the logging intact, which is
desirable for the same reasons as above. ubx.node_rm can still be
called explicitly (i.e. as done by blockdiagram:pulldown) when the
main node is being destructed.

Signed-off-by: Markus Klotzbuecher <mk@mkio.de>