connect: don't allow custom name for non-existing connection targets
This creates headaches for ensuring unique names for hierarchical
sub-configurations. As this non-existing target feature is mainly used
for quick exports of data via mqueues or other iblocks, this shouldn't
be a big issue.

That said, it could be probably be solved with a comparably small
effort by creating the named target at the scope level of the
hierarchical composition.

Signed-off-by: Markus Klotzbuecher <mk@mkio.de>